wip #12: not working but moved code and handlers
