+++ middleware/middleware.go
@@ -1,28 +1,40 @@
-package main
+package middleware
 
 import (
 	"net/http"
 
+	"github.com/getsentry/sugar"
+	"github.com/getsentry/sugar/console"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
 )
 
-func loggingMiddleware(logger *zap.SugaredLogger) gin.HandlerFunc {
+func SetupLoggerMiddleware(useJSON bool) gin.HandlerFunc {
 	return func(c *gin.Context) {
+		// Create a new Sugar logger
+		logger, _ := sugar.New()
 
-		logger.Infow("Request Information",
-			"Method", c.Request.Method,
-			"Path", c.Request.URL.Path,
-			"Query", c.Request.URL.RawQuery,
-			"UserAgent", c.Request.UserAgent(),
-		)
+		if useJSON {
+			// Configure the logger output to use JSON format
+			logger.SetFormatter(sugar.NewJSONFormatter())
+		} else {
+			// Configure the logger output to use the console writer
+			logger.SetWriter(console.New())
+		}
+
+		// You can customize other logger settings here, such as log level, etc.
+		// For example:
+		// logger.SetLevel(sugar.InfoLevel)
+
+		// Attach the logger to the context for access in handlers
 		c.Set("logger", logger)
 
+		// Continue with the request
 		c.Next()
 	}
 }
 
-func authMiddleware(logger *zap.SugaredLogger) gin.HandlerFunc {
+func AuthMiddleware(logger *zap.SugaredLogger) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		// Retrieve the logger from the Gin context
 		c.Set("logger", logger)
